do not expose port 443 because it is not used by default
The default nginx configuration does not configure nginx to be
listening on port 443.

However, the Dockerfile exposes this port, which has some implications;

- users get the expectation that nginx is listening on port 443, but
  when publishing the port discover that no connection can be
  established.
- when using `-P` / `--publish-all`, docker will publish port 443 (using
  random port mapping)
- some proxies use "expose" to detect which ports are provided by a
  container and make configuration decisions based on that information,
  again leading to port 443 being mapped, but not functional.

This change removes port 443 from the Dockerfile.

Removing this port can be a breaking change, so should be documented
(i.e., users should now add `EXPOSE 443` in their Dockerfile if they
extend the nginx image, or use `--expose 443` when running the image.

Signed-off-by: Sebastiaan van Stijn <github@gone.nl>