modules: Added a hack to make build system fetch a requested nginx version
We're now using a tip of default branch on pkg-oss which moves fast and
breaks things.  A better solution is probably to check out a specific
nginx tag and build whatever is available for it.  This, however, is not
currently possible with 1.19.6, and only works for 1.19.7.  This should
probably be reverted once we have enough tags supporting it.