modules: added xslscript.pl to debian-based image.
This should not be really required.  Sometimes when building debian
changelogs in pkg-oss there seems to be a timestamp issue forcing
changes.xslt to be rebuilt.  I can not reliable reproduce it, but a
workaround seems to be just to ship the script and use it when it's
needed.

Fixes #677.