Upgrade openssl-dev package to keep libssl up to date
The current base image, alpine 3.17, does not bundle with the latest version of
libssl and the docker build process does not update the libssl library.
So building the image does not guaranteee to have installed the latest version.
This change take care to build the image using the latest libssl version available
from alpine packages repository.